chore(StoryPage): split some components, add some types
